requirements.txt
requests>=2.28.0
flask>=2.3.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
